Resolve circular dependencies problem in Batch related files.
