+++ src/treemodel/delta/attributedelta.js
@@ -4,9 +4,8 @@
  */
 
 'use strict';
-
 import Delta from './delta.js';
-import register from '../batch.js';
+import { register } from '../batch-base.js';
 import AttributeOperation from '../operation/attributeoperation.js';
 import Position from '../position.js';
 import Range from '../range.js';
