+++ src/treemodel/delta/weakinsertdelta.js
@@ -6,7 +6,7 @@
 'use strict';
 
 import Delta from './delta.js';
-import register from '../batch.js';
+import { register } from '../batch-base.js';
 import InsertOperation from '../operation/insertoperation.js';
 import NodeList from '../nodelist.js';
